refactor(automation): move state managers to state/ subpackage
Extract planner_state, implementer_state, and review_state into hephaestus/automation/state/ subpackage with backward-compatible shims. Consolidates state management into a clear domain boundary with unified imports.

Closes #1443

Implemented-By: claude-opus-4-8
