+++ hephaestus/automation/state/__init__.py
@@ -1,0 +1,7 @@
+"""State-management domain for the automation pipeline.
+
+Canonical homes for the per-issue state modules (planner, implementer,
+review). The original flat module paths
+(``hephaestus.automation.{planner_state,implementer_state,review_state}``)
+remain as thin re-export shims for backward compatibility.
+"""
